+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/CommandContextItem.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class CommandContextItem : ICommandContextItem
+{
+    public bool IsCritical { get; set; }
+    public ICommand Command { get; set; }
+    public string Tooltip { get; set; } = "";
+    public CommandContextItem(ICommand command)
+    {
+        Command = command;
+    }
+}
