+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/DetailsElement.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class DetailsElement : IDetailsElement
+{
+    public string Key { get; set; }
+    public IDetailsData? Data { get; set; }
+}
