+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/DetailsLink.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class DetailsLink : IDetailsLink
+{
+    public Uri Link { get; set; }
+    public string Text { get; set; }
+}
