+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/DetailsTags.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class DetailsTags : IDetailsTags
+{
+    public ITag[] Tags { get; set; }
+}
