+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/DynamicListPage.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class DynamicListPage : ListPage, IDynamicListPage
+{
+    public virtual ISection[] GetItems(string query) => throw new NotImplementedException();
+}
