+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/Filter.cs
@@ -1,0 +1,8 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class Filter : IFilter
+{
+    public IconDataType Icon => throw new NotImplementedException();
+    public string Id => throw new NotImplementedException();
+    public string Name => throw new NotImplementedException();
+}
