+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/FormPage.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class FormPage : Action, IFormPage
+{
+    private bool _Loading = false;
+
+    public bool Loading { get => _Loading; set { _Loading = value; OnPropertyChanged(nameof(Loading)); } }
+
+    public virtual IForm[] Forms() => throw new NotImplementedException();
+}
