+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/GoToPageArgs.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class GoToPageArgs : IGoToPageArgs
+{
+    public required string PageId { get; set; }
+}
