+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/InvokableCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class InvokableCommand : Action, IInvokableCommand
+{
+    public virtual ICommandResult Invoke() => throw new NotImplementedException();
+}
