+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/ListSection.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class ListSection : ISection
+{
+    public string Title { get; set; } = "";
+    public virtual IListItem[] Items { get; set; } = [];
+}
