+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/NoOpAction.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class NoOpAction : InvokableCommand
+{
+    public override ICommandResult Invoke() => ActionResult.KeepOpen();
+}
