+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/SeparatorContextItem.cs
@@ -1,0 +1,5 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class SeparatorContextItem : ISeparatorContextItem
+{
+}
