+++ src/modules/cmdpal/extensionsdk/Microsoft.Windows.CommandPalette.Extensions.Helpers/SeparatorFilterItem.cs
@@ -1,0 +1,5 @@
+﻿namespace Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+public class SeparatorFilterItem : ISeparatorFilterItem
+{
+}
